Rename project to crier
The name 'Sheepdog' was initially a pun, based on knowing the
state of UAT environments (it knows where the ewe-at). However
the implimentation drifted away to something far more generic.
Even with the pun I was a little worried that the title might
be misconstruded, without it its just going to confuse someone.

PDD is great, as long as people get the joke, rather than thinking
they are the butt of it.